revert renaming module
