+++ tests/unit/test_bayesian_optimizer.py
@@ -39,7 +39,7 @@ from trieste.data import Dataset
 from trieste.models import ProbabilisticModel, TrainableProbabilisticModel
 from trieste.observer import OBJECTIVE, Observer
 from trieste.space import Box, SearchSpace
-from trieste.types import TensorType
+from trieste.type import TensorType
 from trieste.utils import Err, Ok
 
 
@@ -310,7 +310,6 @@ def test_bayesian_optimizer_optimize_is_noop_for_zero_steps() -> None:
         .optimize(0, data, {"": _UnusableModel()}, _UnusableRule())
         .astuple()
     )
-
     assert history == []
     final_data = result.unwrap().datasets
     assert len(final_data) == 1
