+++ trieste/acquisition/combination.py
@@ -20,7 +20,7 @@ import tensorflow as tf
 
 from ..data import Dataset
 from ..models import ProbabilisticModel
-from ..types import TensorType
+from ..type import TensorType
 from .function import AcquisitionFunction, AcquisitionFunctionBuilder
