+++ trieste/acquisition/function.py
@@ -29,7 +29,7 @@ import tensorflow_probability as tfp
 from ..data import Dataset
 from ..models import ProbabilisticModel
 from ..space import SearchSpace
-from ..types import TensorType
+from ..type import TensorType
 from ..utils import DEFAULTS
 from ..utils.pareto import Pareto, get_reference_point
 from .sampler import (
