+++ trieste/utils/pareto.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 import tensorflow as tf
 from typing_extensions import Final
 
-from ..types import TensorType
+from ..type import TensorType
 from .misc import DEFAULTS
